Split up editor for Android and iOS.
Preparing changes. Each platform will pick the right one.
